responsive fluid feed posts boxes, styling
This needs to be applied to mobile view. Removing mobile styles and letting the default take over should do it. It's responsive and fluid to fill the viewspace while also increasing the amount of feed post boxes when the view port gets big enough to trigger more boxes to fit in.

Removed category as it makes the content boxes less nice. This way it's sleeker. User will only really be intersted in the category if they like the content and want to find more similar content with the same category, not before actually viewing it.